requirements.txt
polars>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pendulum>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0
pyyaml>=6.0
pyarrow>=15.0.0